Struct crossbeam_deque::Injector

pub struct Injector<T> { /* fields omitted */ }

An injector queue.

This is a FIFO queue that can be shared among multiple threads. Task schedulers typically have a single injector queue, which is the entry point for new tasks.

Examples

use crossbeam_deque::{Injector, Steal};

let q = Injector::new();
q.push(1);
q.push(2);

assert_eq!(q.steal(), Steal::Success(1));
assert_eq!(q.steal(), Steal::Success(2));
assert_eq!(q.steal(), Steal::Empty);

Methods

impl<T> Injector<T>

pub fn new() -> Injector<T>

Creates a new injector queue.

Examples

use crossbeam_deque::Injector;

let q = Injector::<i32>::new();

pub fn push(&self, task: T)

Pushes a task into the queue.

Examples

use crossbeam_deque::Injector;

let w = Injector::new();
w.push(1);
w.push(2);

pub fn steal(&self) -> Steal<T>

Steals a task from the queue.

Examples

use crossbeam_deque::{Injector, Steal};

let q = Injector::new();
q.push(1);
q.push(2);

assert_eq!(q.steal(), Steal::Success(1));
assert_eq!(q.steal(), Steal::Success(2));
assert_eq!(q.steal(), Steal::Empty);

pub fn steal_batch(&self, dest: &Worker<T>) -> Steal<()>

Steals a batch of tasks and pushes them into a worker.

How many tasks exactly will be stolen is not specified. That said, this method will try to steal around half of the tasks in the queue, but also not more than some constant limit.

Examples

use crossbeam_deque::{Injector, Worker};

let q = Injector::new();
q.push(1);
q.push(2);
q.push(3);
q.push(4);

let w = Worker::new_fifo();
q.steal_batch(&w);
assert_eq!(w.pop(), Some(1));
assert_eq!(w.pop(), Some(2));

pub fn steal_batch_and_pop(&self, dest: &Worker<T>) -> Steal<T>

Steals a batch of tasks, pushes them into a worker, and pops a task from that worker.

How many tasks exactly will be stolen is not specified. That said, this method will try to steal around half of the tasks in the queue, but also not more than some constant limit.

Examples

use crossbeam_deque::{Injector, Steal, Worker};

let q = Injector::new();
q.push(1);
q.push(2);
q.push(3);
q.push(4);

let w = Worker::new_fifo();
assert_eq!(q.steal_batch_and_pop(&w), Steal::Success(1));
assert_eq!(w.pop(), Some(2));

pub fn is_empty(&self) -> bool

Returns true if the queue is empty.

Examples

use crossbeam_deque::Injector;

let q = Injector::new();

assert!(q.is_empty());
q.push(1);
assert!(!q.is_empty());