Struct crossbeam_deque::Worker

pub struct Worker<T> { /* fields omitted */ }

The worker side of a deque.

Workers push elements into the back and pop elements depending on the strategy:

In FIFO deques, elements are popped from the front.

In LIFO deques, elements are popped from the back.

A deque has only one worker. Workers are not intended to be shared among multiple threads.

Methods

impl<T> Worker<T>

pub fn is_empty(&self) -> bool

Returns true if the deque is empty.

use crossbeam_deque as deque;

let (w, _) = deque::lifo();
assert!(w.is_empty());
w.push(1);
assert!(!w.is_empty());

pub fn push(&self, value: T)

Pushes an element into the back of the deque.

Examples

use crossbeam_deque as deque;

let (w, _) = deque::lifo();
w.push(1);
w.push(2);

pub fn pop(&self) -> Pop<T>

Pops an element from the deque.

Which end of the deque is used depends on the strategy:

If this is a FIFO deque, an element is popped from the front.

If this is a LIFO deque, an element is popped from the back.

Examples

use crossbeam_deque::{self as deque, Pop};

let (w, _) = deque::fifo();
w.push(1);
w.push(2);

assert_eq!(w.pop(), Pop::Data(1));
assert_eq!(w.pop(), Pop::Data(2));
assert_eq!(w.pop(), Pop::Empty);