Module crossbeam_deque::deque

Structs

Block

A block in a linked list.

Buffer

A buffer that holds tasks in a worker queue.

Injector

An injector queue.

Inner

Internal queue data shared between the worker and stealers.

Position

A position in a queue.

Slot

A slot in a block.

Stealer

A stealer handle of a worker queue.

Worker

A worker queue.

Enums

Flavor

Worker queue flavor: FIFO or LIFO.

Steal

Possible outcomes of a steal operation.

Constants

BLOCK_CAP

DESTROY

FLUSH_THRESHOLD_BYTES

HAS_NEXT

LAP

MAX_BATCH

MIN_CAP

READ

SHIFT

WRITE